merubah route menjadi api resource dan membuat service
